Logos and app description updated
The Dovico App store's logos were updated and I was asked if I could
update the logos in this app to match. When submitted for certification,
it failed because the description that was provided was too generic
(that and I didn't give the testers the amount of info they wanted for
how to create a test Dovico account). This app has now passed
cerification again.